requirements.txt
fastapi
uvicorn
pydantic
faiss-cpu
numpy
sentence-transformers
groq
python-dotenv
